fix(basetype): Java CodeCheck 门禁修复 测试项目 OBJ.05 包装重载消歧 (1/1)
对齐 CI：OBJ.05/CMT.06/FMT.06/FMT.07/FMT.09/FMT.13；assertEqual 包装参数拆箱消歧；本地 java-check 0 违规；javac 通过。
行数约 665。

+++ src/intellij_plugin/ohosgen/src/test/java/basetype/int16array2/Int16ArrayFindTest02.java
@@ -431,7 +431,7 @@ public class Int16ArrayFindTest02 extends BasTest {
     void testInt16ArrayFindTestTwo048() {
     Int16Array arr = Int16Array.of(42, 42, 42, 42);
     Integer result = arr.find((value, index, array) -> true);
-    assertEqual(arr.get(0), result);
+    assertEqual(arr.get(0).intValue(), result);
     }
 
     @Test
